@@ -1,8 +1,8 @@
-using Clean.Architecture.Web.ContributorEndpoints;
+﻿using Clean.Architecture.Web.ContributorEndpoints;
 
 namespace Clean.Architecture.Web.Endpoints.ContributorEndpoints;
 
 public class ContributorListResponse
 {
-  public List<ContributorRecord> Contributors { get; set; } = new();
+  public List<ContributorRecord> Contributors { get; set; } = [];
 }
